refactor: verifyotp を auth モジュールに移動
