+++ app/src/main/java/app/kaito_dogi/mybrary/MybraryNavHost.kt
@@ -9,6 +9,7 @@ import app.kaito_dogi.mybrary.core.navigation.MybraryRoute
 import app.kaito_dogi.mybrary.feature.auth.AuthRoute
 import app.kaito_dogi.mybrary.feature.auth.authNavigation
 import app.kaito_dogi.mybrary.feature.auth.route.login.loginScreen
+import app.kaito_dogi.mybrary.feature.auth.route.verifyotp.verifyOtpScreen
 import app.kaito_dogi.mybrary.feature.mybookdetail.MyBookDetailNavArg
 import app.kaito_dogi.mybrary.feature.mybookdetail.myBookDetailRouteWithNavArg
 import app.kaito_dogi.mybrary.feature.mybookdetail.myBookDetailScreen
@@ -16,7 +17,6 @@ import app.kaito_dogi.mybrary.feature.mybooklist.MyBookListRoute
 import app.kaito_dogi.mybrary.feature.mybooklist.myBookListScreen
 import app.kaito_dogi.mybrary.feature.searchbooks.SearchBooksRoute
 import app.kaito_dogi.mybrary.feature.searchbooks.searchBooksScreen
-import app.kaito_dogi.mybrary.feature.verifyotp.verifyOtpScreen
 
 @Composable
 internal fun MybraryNavHost(
@@ -43,14 +43,14 @@ internal fun MybraryNavHost(
         },
       )
 
+      verifyOtpScreen(
+        onVerifyOtpComplete = {
+          navController.navigate(MyBookListRoute)
+        },
+      )
+
       // TODO: screen の追加
     }
-    // TODO: 他の画面から遷移できないようにする
-    verifyOtpScreen(
-      onVerifyOtpComplete = {
-        navController.navigate(MyBookListRoute)
-      },
-    )
     myBookListScreen(
       onAdditionClick = {
         navController.navigate(SearchBooksRoute)
