--- feature/verify-otp/src/main/java/app/kaito_dogi/mybrary/feature/verifyotp/VerifyOtpNavArg.kt
@@ -1,17 +1,0 @@
-package app.kaito_dogi.mybrary.feature.verifyotp
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-import kotlinx.serialization.Serializable
-
-@Serializable
-@Parcelize
-data class VerifyOtpNavArg(
-  val email: String,
-  val source: Source,
-) : Parcelable {
-  enum class Source {
-    Login,
-    SignUp,
-  }
-}
