+++ settings.gradle.kts
@@ -30,4 +30,3 @@ include(":feature:auth")
 include(":feature:my-book-detail")
 include(":feature:my-book-list")
 include(":feature:search-books")
-include(":feature:verify-otp")
